Move row_reader to internal/
